Move previews out of main.
